+++ mayan/apps/sources/models/base.py
@@ -8,11 +8,11 @@ from django.utils.translation import ugettext_lazy as _
 from django_celery_beat.models import PeriodicTask, IntervalSchedule
 from model_utils.managers import InheritanceManager
 
-from mayan.apps.common.compressed_files import Archive
-from mayan.apps.common.exceptions import NoMIMETypeMatch
 from mayan.apps.converter.layers import layer_saved_transformations
 from mayan.apps.documents.models import Document, DocumentType
 from mayan.apps.documents.settings import setting_language
+from mayan.apps.storage.compressed_files import Archive
+from mayan.apps.storage.exceptions import NoMIMETypeMatch
 
 from ..literals import (
     DEFAULT_INTERVAL, SOURCE_CHOICES, SOURCE_UNCOMPRESS_CHOICES